+++ saleor/graphql/costs_map.py
@@ -1,6 +1,0 @@
-COST_MAP = {
-    "Query": {
-        "productVariant": {"complexity": 1},
-        "productVariants": {"complexity": 1, "multipliers": ["first", "last"]},
-    },
-}