+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Batch loads in queries for Apollo Federation - #8273 by @rafalp
 - Reserve stocks for checkouts - #7589 by @rafalp
 - Add `variant_selection` to `ProductAttributeAssign` operations - #8235 by @kuchichan
+- Add query complexity limit to GraphQL api - #8526 by rafalp
 
 # 3.0.0 [Unreleased]